refactored task class with its static members using pointers and runtime memory allocation as well as used static casts in spi_base class due to compiler warnings

+++ src/spiDevices/SpiBase.cpp
@@ -32,13 +32,13 @@ SpiBase::~SpiBase()
 /// @brief initializes spi-channel
 /// @param sdiPin Serial data in / MISO
 /// @param sdoPin Serial data out / MOSI
-/// @param sclkPin Serial Clock 
+/// @param sclkPin Serial Clock
 /// @param baudrateKhz Baudrate in kHz
 /// @param spiInstance spi instance - ref rpi-pico c/c++ sdk (spi0/spi1)
 void SpiBase::spiInit(uint8_t sdiPin, uint8_t sdoPin, uint8_t sclkPin, uint16_t baudrateKhz, spi_inst_t *spiInstance)
 {
     spi_init(spiInstance, 1000 * baudrateKhz);
-    
+
     gpio_set_function(sdiPin, GPIO_FUNC_SPI);
     gpio_set_function(sdoPin, GPIO_FUNC_SPI);
     gpio_set_function(sclkPin, GPIO_FUNC_SPI);
@@ -56,7 +56,12 @@ void SpiBase::_initCsGpio(uint8_t csPin)
 /// @return register value 32 bit
 uint32_t SpiBase::_spiReadReg(uint8_t reg)
 {
-    uint8_t tx_buffer[5] = {reg & 0x7F, 0, 0, 0, 0}; // Read command - MSB = 0
+    uint8_t tx_buffer[5] = {
+        static_cast<uint8_t>(reg & 0x7F),
+        0,
+        0,
+        0,
+        0}; // Read command - MSB = 0
     uint8_t rx_buffer[5] = {0};
 
     xSemaphoreTake(_spiMutex, pdMS_TO_TICKS(100));
@@ -69,7 +74,7 @@ uint32_t SpiBase::_spiReadReg(uint8_t reg)
 
     xSemaphoreGive(_spiMutex);
 
-    // wait some time 
+    // wait some time
     vTaskDelay(pdMS_TO_TICKS(2));
 
     xSemaphoreTake(_spiMutex, pdMS_TO_TICKS(100));
@@ -96,7 +101,7 @@ uint32_t SpiBase::_spiReadReg(uint8_t reg)
 /// @param reg register address
 /// @param mask register mask fpr specific bit field
 /// @param shift necessary bitshift - refer datasheet
-/// @return 
+/// @return
 uint32_t SpiBase::_spiReadBitField(uint8_t reg, uint32_t mask, uint8_t shift)
 {
     // get Register
@@ -112,11 +117,13 @@ uint32_t SpiBase::_spiReadBitField(uint8_t reg, uint32_t mask, uint8_t shift)
 /// @return true if transmition complete
 bool SpiBase::_spiWriteReg(uint8_t reg, uint32_t data)
 {
-    uint8_t tx_buffer[5] = {(reg | 0x80) & 0xFF, // write - command: MSB = 1
-                            (data >> 24) & 0xFF,
-                            (data >> 16) & 0xFF,
-                            (data >> 8) & 0xFF,
-                            data & 0xFF};
+    uint8_t tx_buffer[5] = {
+        static_cast<uint8_t>((reg | 0x80) & 0xFF), // write - command: MSB = 1
+        static_cast<uint8_t>((data >> 24) & 0xFF),
+        static_cast<uint8_t>((data >> 16) & 0xFF),
+        static_cast<uint8_t>((data >> 8) & 0xFF),
+        static_cast<uint8_t>(data & 0xFF)};
+
     uint8_t rx_buffer[5] = {0};
 
     xSemaphoreTake(_spiMutex, pdMS_TO_TICKS(100));
